tm/requirements.txt
amqp==2.5.0
anyjson==0.3.3
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
dj-database-url==0.5.0
Django==2.2.1
django-cors-headers==3.0.2
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
dnspython==1.16.0
eventlet==0.25.0
greenlet==0.4.15
kombu==4.6.3
monotonic==1.5
psycopg2==2.7.6.1
PyJWT==1.7.1
pytz==2019.1
redis==3.3.2
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
whitenoise==4.1.3
wincertstore==0.2
